Replacing 'Weap id' property key in item panes to avoid duplicate use of the key 'id'. + 1 bugfix (#20)
* MItem.js: duplicate property key used for 'Weap id' and 'Item id'. Solved by adding 'weapon' to items 'hiddenkeys' and filtering 'id' from passing to the weapons mathProperty().

* MItem.js: fixed accidental use of lower() insead of toLowerCase().

* MWpn.js: necessary to stop renderWpnTable from printing 'Weap id' as it is now duplicated. Tested the third arg by forcing true and false and saw no effect. Surmised from comment that this was its original purpose. One issue that I was unable to solve was changing or removing the '[id]' from anything with secondaryeffect.

* MWpn.js: solved an issue where the secondaryeffect was being parsed as an integer representing id instead of as an object.

* MItems.js: Added the ability to search by 'effect' in the items list which returns 'secondaryeffect' || 'secondaryeffectalways'

* MWpn.js: to change the key 'id' to 'secondayreffect~' when appearing in items I used the showName argument as this is only false when coming from an item pane and not a weapon pane.
I renamed hideName back to showName and made it act as an enumerator. 0 will hide the weap id. 1 will show 'id' as normal thoughout the pane as seen on a weapon pane. 2 and 3 will show the appropriate key for 'secondaryeffect' and 'secondaryeffectalways'. Recursive calls to this function will change conditionally depending on the state of showName. I.e. if showName != 1, the appropriate tag will be selected.
MItem.js: change argument to 0 to cooresdpond with change to hideName/showName.

* This file was accidentally left out of the last commit.

* MWpn.js: fixed an issue where the effects ids weren't being displayed properly.

* MItem.js: the 'secondaryeffectalways' key was really long, it caused some properties to take up two lines and pushed the second column to far left. 'weapon' is a much shorter key that applies to both effects.
weapons.csv: it became necessary to add a 'weapon' column to add the key. The column is a duplicate of 'id'. I wasn't sure if it was possible to use 'id' and rename the key to 'weapon' somewhere in code.
MWpn.js: cleaned up effectkeys into one variable and removed the third option for the enumerator 'showName'

* Cleanup and readability
MWpn.js: cleaned up the code. 'showName' is now 'changeKey' which is a better description of what it does. Its also a bool which is a better fit. Renamed 'effectkeys' to 'althiddenkeys' as it is the alternative to hidden keys.
MItem.js: I realised I could move 'weap id' back down closer to its stats where it was in the first place. Also changed the renderWpnTable call to reflect the 3rd argument being a bool again.